style: improve CSS structure and fix minor styling issues
Added padding to the main element for better control of HTML structure and improved maintenance. Fixed CSS errors such as the '/' in .theme-icon and adjusted the width and height of #calcarea for better styling. Added cursor pointer to the switch and fixed the animation of the history button. Also, applied bold text to the calculator title for better readability.